Functions

Examples for BayesianSVD.

BayesianSVD.BayesianSVD

BayesianSVD.Data

BayesianSVD.Pars

BayesianSVD.Posterior

BayesianSVD.PON

BayesianSVD.SampleSVD

BayesianSVD.hpd

BayesianSVD.BayesianSVD — Module

BayesianSVD

Here is my package.

BayesianSVD.Data — Method

Data(Y, ΩU::KernelFunction, ΩV::KernelFunction, k)

Creates the data class.

See also Pars, Posterior, and SampleSVD.

Arguments

Y: data of dimension n × m

ΩU::KernelFunction: Kernel for U matrix, of dimension n × n

ΩV::KernelFunction: Kernel for V matrix, of dimension m × m

k: number of basis functions to keep

To Do

allow for ΩU and ΩV to be of different types (i.e., ΩU is Matern and ΩV is Identity)

Examples

To Do.

BayesianSVD.Pars — Method

Pars(data::Data)

Creates the parameter class of type typeof(data).'

See also Data, Posterior, and SampleSVD.

Arguments

data: data structure of type Data

Examples

To Do.

BayesianSVD.Posterior — Type

Posterior

Structure of type posterior with subtypes Identity, Exponential, Gaussian, or Matern. Contains the raw posterior samples and some means and 95% quantiles of parameters. Plotting associated with the structure.

See also Pars, Data, and SampleSVD.

Examples

Random.seed!(2)

m = 100
n = 100
x = range(-5, 5, n)
t = range(0, 10, m)

ΣU = MaternKernel(x, ρ = 3, ν = 4, metric = Euclidean())
ΣV = MaternKernel(t, ρ = 3, ν = 4, metric = Euclidean())

k = 5
Φ = PON(n, k, ΣU.K)
Ψ = PON(n, k, ΣV.K)

D = diagm([40, 20, 10, 5, 2])

ϵ = rand(Normal(0, sqrt(0.01)), n, m)
Y = Φ * D * Ψ' + ϵ # n × m

ΩU = MaternKernel(x, ρ = 4, ν = 4, metric = Distances.Euclidean())
ΩV = MaternKernel(t, ρ = 4, ν = 4, metric = Distances.Euclidean())
data = Data(Y, ΩU, ΩV, k)
pars = Pars(data)

posterior, pars = SampleSVD(data, pars; nits = 1000, burnin = 500, show_progress = true)

plot(posterior, x, size = (900, 600), basis = 'U', linewidth = 2, c = [:red :green :purple])
plot(posterior, t, size = (900, 500), basis = 'V', linewidth = 2, c = [:red :green :purple])

# for spatial basis functions, provide x and y
plot(posterior, x, y)

BayesianSVD.PON — Method

PON(n, k, Sigma)

Create an n by k orthonormal matrix with covariance matrix Sigma.

Arguments

n: Number of locations

k: Number of basis functions

Sigma: covariance matrix

Examples

m = 100
n = 100
x = range(-5, 5, n)
t = range(0, 10, m)

ΣU = MaternKernel(x, ρ = 3, ν = 4, metric = Euclidean())
ΣV = MaternKernel(t, ρ = 3, ν = 4, metric = Euclidean())

k = 5
Φ = PON(n, k, ΣU.K)
Ψ = PON(n, k, ΣV.K)

BayesianSVD.SampleSVD — Method

SampleSVD(data::Data, pars::Pars; nits = 10000, burnin = 5000, show_progress = true)

Runs the MCMC sampler for the Bayesian SVD model.

See also Pars, Data, and Posterior.

Arguments

data: Data structure of type Identity, Exponential, Gaussian, or Matern

pars: Parameter structure of type Identity, Exponential, Gaussian, or Matern

Optional Arguments

nits = 10000: Total number of posterior samples to compute

burnin = 5000: Number of samples discarded as burnin

show_progress = true: Indicator on whether to show a progress bar (true) or not (false).

Examples

Random.seed!(2)

m = 100
n = 100
x = range(-5, 5, n)
t = range(0, 10, m)

ΣU = MaternKernel(x, ρ = 3, ν = 4, metric = Euclidean())
ΣV = MaternKernel(t, ρ = 3, ν = 4, metric = Euclidean())

k = 5
Φ = PON(n, k, ΣU.K)
Ψ = PON(n, k, ΣV.K)

D = diagm([40, 20, 10, 5, 2])

ϵ = rand(Normal(0, sqrt(0.01)), n, m)
Y = Φ * D * Ψ' + ϵ # n × m

ΩU = MaternKernel(x, ρ = 4, ν = 4, metric = Distances.Euclidean())
ΩV = MaternKernel(t, ρ = 4, ν = 4, metric = Distances.Euclidean())
data = Data(Y, ΩU, ΩV, k)
pars = Pars(data)

posterior, pars = SampleSVD(data, pars; nits = 1000, burnin = 500, show_progress = true)

BayesianSVD.hpd — Method

hpd(x; p=0.95)

Computes the highest posterior density interval of x.

Arguments

x: vector of data.

Optional Arguments

p: value between 0 and 1 for the probability level.

Examples

x = [1, 3, 2, 5, .2, 1,9]
hpd(x)